requirements.txt
django-bootstrap4
django_taggit